+++ tests/jest/module/_template0_camera-environment-map.ts
@@ -13,10 +13,7 @@ setCameraSource("face");
 const envMap = new ZapparThree.CameraEnvironmentMap();
 scene.environment = envMap.environmentMap;
 
-const sphere = new THREE.Mesh(
-  new THREE.SphereBufferGeometry(1, 16, 12),
-  new THREE.MeshStandardMaterial({ metalness: 1.0, roughness: 0.0 })
-);
+const sphere = new THREE.Mesh(new THREE.SphereGeometry(1, 16, 12), new THREE.MeshStandardMaterial({ metalness: 1.0, roughness: 0.0 }));
 
 sphere.position.set(0, 0, -5);
 scene.add(sphere);